Add correct output for several demos that fail.
For some reason that I'm unable to determine, certain demos exit with a
V_DrawPatch error when they are played back using statdump.exe and
the -nodraw parameter. Because of this, the statistics information is
not recorded. I have manually recorded the correct output from these
demos to fix the problem.